feat(nav): 3-route IA — Chat · My Account · Admin (5 sub-tabs) (#145, #146, #147, #148, #149)
Top-level navigation restructured per Epic #144. Replaces 5 admin
top-level pages + admin-only sub-tabs of User Settings with a single
🛠️ Admin umbrella containing 5 cleanly-named sub-tabs. Cuts 4
duplicated charts on the way through.

Top-level navigation (#145):
- app.py now registers exactly 2 non-admin pages (🤖 Chat, 👤 My
  Account) + 1 admin-gated 🛠️ Admin umbrella. The legacy User
  Settings / Admin Analytics / Feedback Dashboard / Agentic
  Analytics top-level entries are gone.
- New views/my_account.py renders the Epic #124 Profile + Change
  Password forms plus a caption pointing at the sidebar ⚙️ Settings
  dialog for display preferences.
- New views/admin.py is the umbrella page: _guard_admin() at the
  top, a single Time Range segmented control + Refresh Data button
  drives days_int into each sub-tab, then 5 st.tabs in the spec
  order: Users · Training · Analytics · Audit · Feedback.

Split views/user.py (#146):
- views/admin_users.py:render(days_int) hosts the post-Epic #129
  Manage Users layout (toolbar + L/R split + 4 inner tabs).
- views/admin_training.py:render(days_int) hosts the post-Epic #126
  Training Data library-first layout.
- New shared views/_admin_helpers.py holds the @st.dialog helpers
  (pop_train, confirm_destructive, create_user_dialog,
  bulk_import_dialog, export_users_dialog, set_password_dialog,
  _delete_and_rerun) + import_users / export_users_to_csv /
  delete_all_training / export_training_data_to_csv /
  import_training_data_from_csv utilities.
- views/user.py deleted.

Admin → Analytics consolidation + chart cuts (#147):
- views/admin_analytics_consolidated.py:render(days_int) dispatches
  to 4 inner views: Overview · LLM Performance · Errors · Agentic.
  (User Activity moves to the Audit sub-tab — see #148.)
- views/admin_agentic.py absorbs the former
  views/agent_analytics.py:main() content (Runs / Tools / Patient
  Access inner tabs) as a render(days_int) entrypoint.
- Chart cuts applied per Epic AC:
  · Overview "Top Users by Questions" bar chart (removed)
  · Overview "All Users Stats" table (removed)
  · User Activity "Activity Type Breakdown" pie chart (removed,
    counts table preserved)
  · Errors "by Severity" pie chart (removed, counts table preserved)
- views/agent_analytics.py deleted.

Admin → Audit consolidation + deep-link re-target (#148):
- views/admin_audit.py:render(days_int) dispatches to 3 inner views:
  Questions (was Audit Trail) · Admin Actions · User Activity.
- 3 cross-Epic deep-link targets re-pointed from
  st.switch_page("views/admin_analytics.py") to
  st.switch_page("views/admin.py"):
  · views/chat_bot.py Question History accordion "View all in
    audit →" link.
  · views/admin_users.py Activity inner tab "View question audit
    for <username> →" button.
- The PR #151 JS shim on the Overview tab's "View full Audit
  Trail →" link now targets the outer Admin tab labelled "Audit"
  instead of the old "Audit Trail" inner tab.
- audit_trail_pref_user_id session-state contract preserved verbatim.

Admin → Feedback relocation (#149):
- views/admin_feedback.py's main() renamed to render(days_int).
  st.title("Feedback Dashboard") dropped (the Admin tab label
  provides the heading). Every internal helper preserved verbatim.

Smoke-verified end-to-end:
- Admin sidebar shows exactly 3 pages: Chat, My Account, Admin.
- /admin renders 5 outer tabs in spec order: Users, Training,
  Analytics, Audit, Feedback. All inner tabs render.
- /my_account shows "My Account" title, Profile + Change Password
  sections, and the caption pointing at the sidebar Settings
  dialog.
- 157 orm tests passing.

Bookmark breakage to /admin_analytics, /admin_feedback,
/agent_analytics, /user is accepted per Epic Out of Scope.
