--- views/user.py
@@ -1,1009 +1,0 @@
-import io
-
-import pandas as pd
-import streamlit as st
-from pandas import DataFrame
-
-from orm.functions import (
-    admin_change_password,
-    change_password,
-    create_user,
-    delete_all_messages,
-    delete_user,
-    get_all_user_roles,
-    get_all_users,
-    get_user_daily_stats,
-    get_user_recent_questions,
-    get_user_stats_for_all_users,
-    get_users_for_export,
-    update_user,
-    update_user_preferences,
-)
-from orm.models import RoleTypeEnum, SessionLocal, User, UserRole
-from utils.authentication_management import get_user_list_excel
-from utils.chat_bot_helper import get_vn
-from utils.enums import ThemeType, user_selectable_themes
-from utils.vanna_calls import auto_generate_sql_pairs, refresh_stats, train_all
-
-# Get the current user ID from session state cookies. Guarded so the module
-# stays importable when no Streamlit runtime is attached (e.g., pytest);
-# mirrors the defensive pattern in orm.functions._get_current_user_id.
-try:
-    user_id = st.session_state.cookies.get("user_id")
-except Exception:
-    user_id = None
-# Get the current user role from session state (not cookies) and default to least privileged
-try:
-    user_role = st.session_state.get("user_role", RoleTypeEnum.PATIENT.value)
-except Exception:
-    user_role = RoleTypeEnum.PATIENT.value
-# Don't get training data at module load time - get it when rendering the page
-# df = vn.get_training_data()
-
-from utils.quick_logger import get_logger, pvlog
-
-try:
-    pvlog("debug", f"{st.session_state.to_dict()=}")
-except Exception:
-    pass
-
-logger = get_logger(__name__)
-
-
-def _cell_text(raw) -> str:
-    """Return the cleaned text for a spreadsheet cell.
-
-    pandas reads blank Excel cells as float NaN, which str()-ifies to 'nan'
-    (truthy) and would silently bypass the empty-field pre-flight checks.
-    Returns "" when the cell is missing, NaN, or whitespace-only.
-    """
-    if raw is None:
-        return ""
-    try:
-        if pd.isna(raw):
-            return ""
-    except (TypeError, ValueError):
-        pass
-    text = str(raw).strip()
-    return "" if text.lower() == "nan" else text
-
-
-def import_users():
-    """
-    Import users from Excel file and save them to the database.
-
-    Expected columns: UserID, start_password, First Name, Last Name, Email, Organization.
-    Both `Email` / `email` and `Organization` / `organization` casings are accepted.
-
-    Rows with missing or invalid fields are skipped and surfaced in the
-    failures expander. Validation + persistence is delegated to
-    `orm.functions.create_user`, which enforces email format, email
-    uniqueness (case-insensitive), and organization presence.
-    """
-    try:
-        import os
-
-        # Use robust path resolution
-        possible_paths = [
-            "./utils/config/user_list.xlsx",
-            "utils/config/user_list.xlsx",
-            os.path.join(os.path.dirname(__file__), "..", "utils", "config", "user_list.xlsx"),
-        ]
-
-        file_path = None
-        for path in possible_paths:
-            abs_path = os.path.abspath(path)
-            if os.path.exists(abs_path):
-                file_path = abs_path
-                break
-
-        if not file_path:
-            st.error("Excel file not found. Please ensure 'utils/config/user_list.xlsx' exists.")
-            return False
-
-        # Read the Excel file (with CSV fallback)
-        try:
-            df = get_user_list_excel(file_path)
-        except:
-            # Try CSV fallback
-            csv_path = file_path.replace(".xlsx", ".csv")
-            if os.path.exists(csv_path):
-                st.info("Using CSV fallback file")
-                df = pd.read_csv(csv_path)
-            else:
-                st.error("Excel file requires openpyxl library. Please install openpyxl or provide a CSV file.")
-                return False
-
-        if df is False or df is None:
-            st.error(f"Failed to read user list Excel file. Please check the file format and dependencies.")
-            return False
-
-        if df.empty:
-            st.warning("Excel file contains no data")
-            return False
-
-        # Track import results
-        success_count = 0
-        failed_count = 0
-        failed_users = []
-
-        # Look up roles once (read-only) so the per-row create_user calls don't
-        # need their own UserRole query.
-        with SessionLocal() as session:
-            roles = session.query(UserRole).all()
-            role_map = {role.role_name.lower(): role.id for role in roles}
-            role_map.update(
-                {
-                    "administrator": role_map.get("admin", None),
-                    "physician": role_map.get("doctor", None),
-                }
-            )
-            _patient_fallback_role_id = role_map.get("patient", 1)
-
-        for index, row in df.iterrows():
-            try:
-                username = _cell_text(row.get("UserID"))
-                password = _cell_text(row.get("start_password"))
-                first_name = _cell_text(row.get("First Name "))  # trailing space
-                last_name = _cell_text(row.get("Last Name "))  # trailing space
-                email = _cell_text(row.get("Email", row.get("email")))
-                organization = _cell_text(row.get("Organization", row.get("organization")))
-                role_name = "Doctor"  # Default role since not in Excel
-
-                # Skip rows with missing required data
-                if not username or not password:
-                    failed_count += 1
-                    failed_users.append(f"Row {index + 2}: Missing username or password")
-                    continue
-                if not email:
-                    failed_count += 1
-                    failed_users.append(f"Row {index + 2} ({username}): Missing email")
-                    continue
-                if not organization:
-                    failed_count += 1
-                    failed_users.append(f"Row {index + 2} ({username}): Missing organization")
-                    continue
-
-                # Map role name to role ID
-                role_id = role_map.get(role_name.lower())
-                if not role_id:
-                    role_id = _patient_fallback_role_id
-                    logger.warning(f"Role '{role_name}' not found for user {username}, defaulting to Patient")
-
-                # Delegate format + uniqueness validation and persistence to
-                # create_user — single source of truth per Epic #98.
-                ok = create_user(
-                    username,
-                    password,
-                    first_name if first_name else username,
-                    last_name if last_name else "",
-                    role_id,
-                    email=email,
-                    organization=organization,
-                )
-                if ok:
-                    success_count += 1
-                else:
-                    failed_count += 1
-                    failed_users.append(
-                        f"{username}: create_user rejected (duplicate username/email, bad email format, or invalid input)"
-                    )
-
-            except Exception as e:
-                failed_count += 1
-                failed_users.append(f"{row.get('username', f'Row {index + 2}')}: {str(e)}")
-                logger.error(f"Error importing user at row {index + 2}: {e}")
-
-        # Display results
-        if success_count > 0:
-            st.success(f"✅ Successfully imported {success_count} user(s)")
-
-        if failed_count > 0:
-            st.warning(f"⚠️ Failed to import {failed_count} user(s)")
-            if failed_users:
-                with st.expander("Failed imports details"):
-                    for failure in failed_users:
-                        st.text(f"• {failure}")
-
-        return success_count > 0
-
-    except Exception as e:
-        st.error(f"Error during user import: {e}")
-        logger.error(f"Error during user import: {e}")
-        return False
-
-
-def delete_all_training():
-    try:
-        # Get training data with role-based filtering - users can only delete what they can see
-        vn = get_vn()
-        training_data = vn.get_training_data()
-        for index, row in training_data.iterrows():
-            vn.remove_from_training(row["id"])
-        st.toast("Training Data Deleted Successfully!")
-    except Exception as e:
-        st.error(f"An error occurred: {e}")
-        logger.error(f"An error occurred: {e}")
-
-
-def export_training_data_to_csv():
-    try:
-        # Get training data with current user's role-based filtering
-        vn = get_vn()
-        training_data = vn.get_training_data()
-
-        if isinstance(training_data, DataFrame) and not training_data.empty:
-            # Create CSV buffer
-            csv_buffer = io.StringIO()
-            training_data.to_csv(csv_buffer, index=False)
-            csv_data = csv_buffer.getvalue()
-
-            # Generate filename with timestamp
-            import datetime
-
-            timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-            filename = f"training_data_{timestamp}.csv"
-
-            # Provide download button
-            st.download_button(
-                label="📥 Download CSV",
-                data=csv_data,
-                file_name=filename,
-                mime="text/csv",
-                help="Download training data as CSV file",
-            )
-            st.toast("CSV export ready for download!")
-        else:
-            st.warning("No training data available to export.")
-    except Exception as e:
-        st.error(f"An error occurred during CSV export: {e}")
-        logger.error(f"CSV export error: {e}")
-
-
-def export_users_to_csv():
-    try:
-        try:
-            admin_id_raw = st.session_state.cookies.get("user_id")
-        except Exception:
-            admin_id_raw = None
-        if not admin_id_raw:
-            st.error("Could not determine current user.")
-            return
-        admin_id = int(admin_id_raw)
-
-        rows, n = get_users_for_export(admin_id)
-
-        if n == 0:
-            st.warning("No users available to export.")
-            return
-
-        df = DataFrame(rows)
-        csv_buffer = io.StringIO()
-        df.to_csv(csv_buffer, index=False)
-        csv_data = csv_buffer.getvalue()
-
-        import datetime
-
-        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-        filename = f"users_{timestamp}.csv"
-
-        st.download_button(
-            label="📥 Download CSV",
-            data=csv_data,
-            file_name=filename,
-            mime="text/csv",
-            help="Download users as CSV file",
-        )
-    except Exception as e:
-        st.error(f"An error occurred during CSV export: {e}")
-        logger.error(f"User export error: {e}")
-
-
-def import_training_data_from_csv(uploaded_file):
-    try:
-        # Read the uploaded CSV file
-        df = pd.read_csv(uploaded_file)
-
-        # Validate required columns
-        required_columns = ["training_data_type", "question", "content"]
-        missing_columns = [col for col in required_columns if col not in df.columns]
-
-        if missing_columns:
-            st.error(f"Missing required columns: {', '.join(missing_columns)}")
-            return
-
-        # Get Vanna instance
-        vn = get_vn()
-        success_count = 0
-        error_count = 0
-
-        # Process each row
-        progress_bar = st.progress(0)
-        total_rows = len(df)
-
-        for index, row in df.iterrows():
-            try:
-                # Update progress
-                progress_bar.progress((index + 1) / total_rows)
-
-                # Skip rows with missing essential data
-                if pd.isna(row["content"]) or str(row["content"]).strip() == "":
-                    error_count += 1
-                    continue
-
-                training_type = str(row["training_data_type"]).lower()
-                question = str(row["question"]) if not pd.isna(row["question"]) else None
-                content = str(row["content"])
-
-                # Train based on the data type
-                if training_type == "sql" and question and question.strip():
-                    # SQL training data
-                    vn.train(question=question, sql=content)
-                elif training_type in ["ddl", "documentation"]:
-                    # Documentation or DDL training data
-                    vn.train(documentation=content)
-                else:
-                    # Default to documentation if unclear
-                    vn.train(documentation=content)
-
-                success_count += 1
-
-            except Exception as row_error:
-                error_count += 1
-                logger.error(f"Error processing row {index}: {row_error}")
-
-        progress_bar.empty()
-
-        # Show results
-        if success_count > 0:
-            st.success(f"Successfully imported {success_count} training entries!")
-            if error_count > 0:
-                st.warning(f"{error_count} entries failed to import. Check logs for details.")
-            st.rerun()
-        else:
-            st.error("No training data was successfully imported.")
-
-    except Exception as e:
-        st.error(f"An error occurred during CSV import: {e}")
-        logger.error(f"CSV import error: {e}")
-
-
-@st.dialog("Cast your vote")
-def pop_train(type):
-    try:
-        from utils.security_validator import security_validator
-
-        if type == "sql":
-            with st.form("add_training_data"):
-                question = st.text_input("Question")
-                content1 = st.text_input("Sql")
-                if st.form_submit_button("Add"):
-                    if question == "" or content1 == "":
-                        st.error("Please fill all fields.")
-                    else:
-                        is_valid, violations = security_validator.validate_sql_content(content1)
-                        if not is_valid:
-                            st.error(f"SQL contains forbidden references: {', '.join(violations)}")
-                        else:
-                            get_vn().train(question=question, sql=content1)
-                            st.toast("Training Data Added Successfully!")
-                            st.rerun()
-        else:
-            with st.form("add_training_data"):
-                content2 = st.text_input("Documentation")
-                if st.form_submit_button("Add"):
-                    if content2 == "":
-                        st.error("Please fill all fields.")
-                    else:
-                        is_valid, violations = security_validator.validate_documentation(content2)
-                        if not is_valid:
-                            st.error(f"Documentation contains forbidden references: {', '.join(violations)}")
-                        else:
-                            get_vn().train(documentation=content2)
-                            st.toast("Training Data Added Successfully!")
-                            st.rerun()
-    except Exception as e:
-        st.error(f"An error occurred: {e}")
-        logger.error(f"An error occurred: {e}")
-
-
-@st.dialog("Confirm destructive action")
-def confirm_destructive(body_md: str, token: str, on_confirm, *, button_label: str = "Confirm"):
-    st.markdown(body_md)
-    st.warning("This action cannot be undone.")
-    typed = st.text_input(f"Type `{token}` to confirm:", key=f"_confirm_typed_{button_label}")
-    cols = st.columns(2)
-    with cols[0]:
-        if st.button("Cancel", key=f"_confirm_cancel_{button_label}"):
-            st.rerun()
-    with cols[1]:
-        if st.button(
-            button_label,
-            type="primary",
-            disabled=(typed != token),
-            key=f"_confirm_go_{button_label}",
-        ):
-            on_confirm()
-            st.rerun()
-
-
-@st.dialog("Create User")
-def create_user_dialog():
-    roles = get_all_user_roles()
-    role_id_by_name = {name: rid for rid, name, _ in roles}
-    role_names = [name for rid, name, _ in roles]
-    theme_options = user_selectable_themes()
-    with st.form("create_user_dialog_form", clear_on_submit=True):
-        cu_username = st.text_input("Username")
-        cu_password = st.text_input("Temporary Password", type="password")
-        cu_first = st.text_input("First Name")
-        cu_last = st.text_input("Last Name")
-        cu_email = st.text_input("Email")
-        cu_organization = st.text_input("Organization")
-        cu_role_name = st.selectbox(
-            "Role", options=role_names, index=role_names.index("Patient") if "Patient" in role_names else 0
-        )
-        cu_theme = st.selectbox("Theme", options=theme_options, index=0)
-        submitted = st.form_submit_button("Create User", type="primary")
-        if submitted:
-            if (
-                not cu_username
-                or not cu_password
-                or not cu_first
-                or not cu_email.strip()
-                or not cu_organization.strip()
-            ):
-                st.error("Please provide username, password, first name, email, and organization.")
-            else:
-                ok = create_user(
-                    cu_username,
-                    cu_password,
-                    cu_first,
-                    cu_last,
-                    role_id_by_name.get(cu_role_name),
-                    email=cu_email,
-                    organization=cu_organization,
-                    theme=cu_theme,
-                )
-                if ok:
-                    st.success("User created.")
-                    st.rerun()
-                else:
-                    st.error(
-                        "Failed to create user. Possible causes: username or email already exists, "
-                        "email format is invalid."
-                    )
-
-
-@st.dialog("Bulk Import Users")
-def bulk_import_dialog():
-    st.info(
-        "📁 Import users from Excel file: `./utils/config/user_list.xlsx`. "
-        "Required columns: UserID, start_password, First Name, Last Name, Email, Organization."
-    )
-    bi_col1, bi_col2 = st.columns(2)
-    with bi_col1:
-        if st.button("Import Users", type="primary", key="bulk_import_dialog_run"):
-            import_users()
-    with bi_col2:
-        if st.button("Close", key="bulk_import_dialog_close"):
-            st.rerun()
-
-
-@st.dialog("Export Users")
-def export_users_dialog():
-    st.info(
-        "📤 Download all users as CSV. Columns match Bulk User Import "
-        "(UserID, First Name, Last Name, Email, Organization, Role). "
-        "Passwords are not included."
-    )
-    export_users_to_csv()
-    if st.button("Close", key="export_users_dialog_close"):
-        st.rerun()
-
-
-@st.dialog("Set Password")
-def set_password_dialog(selected: dict):
-    st.markdown(f"Set a new password for **{selected['username']}**.")
-    new_pw = st.text_input("New Password", type="password", key="set_pw_dialog_pw")
-    sp_col1, sp_col2 = st.columns(2)
-    with sp_col1:
-        if st.button("Set Password", type="primary", key="set_pw_dialog_submit"):
-            if not new_pw:
-                st.error("Enter a new password.")
-            elif admin_change_password(selected["id"], new_pw):
-                st.success("Password updated.")
-                st.rerun()
-            else:
-                st.error("Failed to update password.")
-    with sp_col2:
-        if st.button("Cancel", key="set_pw_dialog_cancel"):
-            st.rerun()
-
-
-def _delete_and_rerun(user_id: int):
-    if delete_user(user_id):
-        st.toast("User deleted.")
-        st.rerun()
-    else:
-        st.error("Failed to delete user.")
-
-
-st.title("User Settings")
-
-tabs = ["My Account"]
-if st.session_state.get("user_role") == RoleTypeEnum.ADMIN.value:
-    tabs = ["My Account", "Training Data", "Manage Users"]
-tab_objects = st.tabs(tabs)
-tab1 = tab_objects[0]
-tab2 = tab_objects[1] if len(tab_objects) > 1 else None
-tab3 = tab_objects[2] if len(tab_objects) > 2 else None
-
-with tab1:
-    # My Profile — every logged-in user can view and update their own email
-    # and organization. The admin Edit User panel in tab3 covers admin
-    # edits-on-behalf; this section is for self-service.
-    if user_id:
-        try:
-            with SessionLocal() as session:
-                _me = session.query(User).filter(User.id == int(user_id)).first()
-        except Exception:
-            _me = None
-        if _me is not None:
-            st.markdown("**Profile**")
-            with st.form("my_profile_form"):
-                my_email = st.text_input("Email", value=_me.email or "")
-                my_organization = st.text_input("Organization", value=_me.organization or "")
-                if st.form_submit_button("Save Profile", type="primary"):
-                    ok = update_user(
-                        int(user_id),
-                        email=my_email.strip() or None,
-                        organization=my_organization.strip() or None,
-                    )
-                    if ok:
-                        st.success("Profile updated.")
-                        st.rerun()
-                    else:
-                        st.error(
-                            "Failed to update profile. Possible causes: email already "
-                            "in use, email format is invalid, or organization is empty."
-                        )
-            st.divider()
-
-    st.markdown("**Change Password**")
-    with st.form("change_password_form"):
-        current_password = st.text_input("Current Password", type="password")
-        new_password = st.text_input("New Password", type="password")
-        confirm_new_password = st.text_input("Confirm New Password", type="password")
-        submit_button = st.form_submit_button("Change Password", type="primary")
-
-        if submit_button:
-            if new_password != confirm_new_password:
-                st.error("New password and confirmation do not match.")
-            else:
-                if change_password(user_id, current_password, new_password):
-                    st.success("Password changed successfully.")
-                else:
-                    st.error("Current password is incorrect.")
-
-if tab2 and st.session_state.get("user_role") == RoleTypeEnum.ADMIN.value:
-    with tab2:
-        # Section 1 — Training Data Library
-        st.subheader("Training Data Library")
-        df = get_vn().get_training_data()
-        if isinstance(df, DataFrame) and not df.empty:
-            display_df = df[["id", "training_data_type", "question", "content"]].copy()
-            display_df.columns = ["ID", "Type", "Question", "Content"]
-
-            filter_col1, filter_col2 = st.columns([0.6, 0.4])
-            with filter_col1:
-                search_query = st.text_input(
-                    "Search training data",
-                    placeholder="Search by question or content...",
-                    key="training_search",
-                )
-            with filter_col2:
-                available_types = sorted(display_df["Type"].dropna().unique().tolist())
-                type_filter = st.multiselect("Filter by type", options=available_types, key="training_type_filter")
-
-            if search_query:
-                mask = display_df["Question"].astype(str).str.contains(search_query, case=False, na=False) | display_df[
-                    "Content"
-                ].astype(str).str.contains(search_query, case=False, na=False)
-                display_df = display_df[mask]
-            if type_filter:
-                display_df = display_df[display_df["Type"].isin(type_filter)]
-
-            if st.session_state.cookies.get("role_name") == "Admin":
-                display_df.insert(0, "Select", False)
-
-                edited_df = st.data_editor(
-                    display_df,
-                    use_container_width=True,
-                    hide_index=True,
-                    column_config={
-                        "Select": st.column_config.CheckboxColumn(
-                            "Delete?",
-                            help="Select rows to delete",
-                            default=False,
-                            width="small",
-                        ),
-                        "ID": st.column_config.TextColumn(
-                            "ID",
-                            width="small",
-                            disabled=True,
-                        ),
-                        "Type": st.column_config.TextColumn(
-                            "Type",
-                            width="small",
-                            disabled=True,
-                        ),
-                        "Question": st.column_config.TextColumn(
-                            "Question",
-                            width="medium",
-                            disabled=True,
-                        ),
-                        "Content": st.column_config.TextColumn(
-                            "Content",
-                            width="large",
-                            disabled=True,
-                        ),
-                    },
-                    num_rows="fixed",
-                    key="training_data_grid",
-                )
-
-                selected_rows = edited_df[edited_df["Select"]]
-                if not selected_rows.empty:
-                    col1, col2 = st.columns([1, 4])
-                    with col1:
-                        if st.button(
-                            f"🗑️ Delete {len(selected_rows)} Selected",
-                            type="primary",
-                            help="Delete all selected training data entries",
-                        ):
-                            vn = get_vn()
-                            for _, row in selected_rows.iterrows():
-                                vn.remove_from_training(row["ID"])
-                            st.toast(f"Deleted {len(selected_rows)} training data entries!")
-                            st.rerun()
-                    with col2:
-                        st.caption(f"Selected {len(selected_rows)} of {len(display_df)} entries for deletion")
-            else:
-                st.dataframe(
-                    display_df[["ID", "Type", "Question", "Content"]],
-                    use_container_width=True,
-                    hide_index=True,
-                    column_config={
-                        "ID": st.column_config.TextColumn("ID", width="small"),
-                        "Type": st.column_config.TextColumn("Type", width="small"),
-                        "Question": st.column_config.TextColumn("Question", width="medium"),
-                        "Content": st.column_config.TextColumn("Content", width="large"),
-                    },
-                )
-
-            st.caption(f"Total: {len(df)} training data entries")
-        else:
-            st.info("No training data available.")
-
-        # Section 2 — Add Training Data
-        st.divider()
-        st.subheader("Add Training Data")
-        add_col1, add_col2 = st.columns(2)
-        with add_col1:
-            if st.button("Add SQL", key="add_sql_btn"):
-                pop_train("sql")
-        with add_col2:
-            if st.button("Add Documentation", key="add_doc_btn"):
-                pop_train("documentation")
-
-        # Section 3 — Bulk Operations
-        st.divider()
-        st.subheader("Bulk Operations")
-
-        st.markdown("**Import CSV**")
-        uploaded_file = st.file_uploader(
-            "Choose a CSV file to upload training data",
-            type=["csv"],
-            help="Upload a CSV file with columns: training_data_type, question, content",
-        )
-        if uploaded_file is not None:
-            st.info(f"📁 File: {uploaded_file.name} ({uploaded_file.size} bytes)")
-            imp_col1, imp_col2 = st.columns(2)
-            with imp_col1:
-                if st.button("📤 Import Training Data", type="primary"):
-                    with st.spinner("Importing training data..."):
-                        import_training_data_from_csv(uploaded_file)
-            with imp_col2:
-                if st.button("❌ Cancel"):
-                    st.rerun()
-
-        st.markdown("**Export CSV**")
-        if st.button("Export CSV", key="export_csv_btn"):
-            export_training_data_to_csv()
-
-        st.markdown("**Auto-Generate SQL Pairs**")
-        st.caption("Use the connected LLM to generate, validate, and train SQL question-answer pairs automatically.")
-        gen_col1, gen_col2, gen_spacer = st.columns((0.20, 0.20, 0.60))
-        with gen_col1:
-            pair_count = st.number_input(
-                "Number of pairs",
-                min_value=1,
-                max_value=50,
-                value=5,
-                step=1,
-                key="auto_gen_pair_count",
-            )
-        with gen_col2:
-            st.write("")  # vertical spacing
-            st.write("")
-            if st.button("Generate", type="primary", key="auto_gen_btn"):
-                with st.status("Generating SQL training pairs...", expanded=True) as gen_status:
-                    gen_results = auto_generate_sql_pairs(count=int(pair_count))
-                    error_msg = gen_results.get("error")
-                    if error_msg:
-                        gen_status.update(
-                            label=f"Generation failed: {error_msg}",
-                            state="error",
-                        )
-                        st.error(error_msg)
-                    else:
-                        passed = gen_results.get("passed", 0)
-                        failed = gen_results.get("failed", 0)
-                        if passed > 0:
-                            gen_status.update(
-                                label=f"Generated {passed} pair(s) ({failed} failed)",
-                                state="complete",
-                            )
-                            st.success(f"Successfully trained {passed} SQL pair(s).")
-                        else:
-                            gen_status.update(
-                                label=f"Generation complete: 0 passed, {failed} failed",
-                                state="error",
-                            )
-                            st.warning("No valid pairs were generated. Check LLM connectivity and training data.")
-                    failed_details = [d for d in gen_results.get("details", []) if d["status"] == "failed"]
-                    if failed_details:
-                        with st.expander(f"Failed pairs ({len(failed_details)})"):
-                            for d in failed_details:
-                                st.text(f"Pair {d['index']}: {d.get('reason', 'Unknown')}")
-
-        # Section 4 — Training Pipeline
-        st.divider()
-        st.subheader("Training Pipeline")
-        train_col1, train_col2, _train_spacer = st.columns((0.20, 0.20, 0.60))
-        with train_col1:
-            st.button(
-                "Train All",
-                type="primary",
-                on_click=train_all,
-                help="Full pipeline: DDL, Schema Plan, Auto-Enhance, AI Docs",
-            )
-        with train_col2:
-            st.button(
-                "Refresh Stats",
-                on_click=refresh_stats,
-                help="Re-run column statistics only (use when data changed but schema is the same)",
-            )
-
-        # Section 5 — Danger Zone
-        st.divider()
-        st.subheader("Danger Zone")
-        if st.button("Remove All Training Data", key="danger_remove_all_btn"):
-            confirm_destructive(
-                body_md=(
-                    "Permanently removes every training-data row visible to you "
-                    "(role-filtered). Schema embeddings will need to be "
-                    "regenerated via **Train All**."
-                ),
-                token="DELETE",
-                on_confirm=delete_all_training,
-                button_label="Remove All Training Data",
-            )
-        if st.button("Delete My Chat History", key="danger_delete_chat_btn"):
-            confirm_destructive(
-                body_md=(
-                    "Permanently removes **your** chat history from this account. "
-                    "Other users' messages are not affected."
-                ),
-                token="DELETE",
-                on_confirm=delete_all_messages,
-                button_label="Delete My Chat History",
-            )
-
-if tab3 and st.session_state.get("user_role") == RoleTypeEnum.ADMIN.value:
-    with tab3:
-        st.subheader("Manage Users")
-        st.caption("Create, edit, or remove users. View settings and activity stats.")
-
-        roles = get_all_user_roles()
-        role_id_by_name = {name: rid for rid, name, _ in roles}
-        role_names = [name for rid, name, _ in roles]
-
-        stats_map = get_user_stats_for_all_users()
-
-        # Toolbar — admin-action dialogs
-        tb1, tb2, tb3_col, _spacer = st.columns([0.18, 0.18, 0.18, 0.46])
-        with tb1:
-            if st.button("+ Create User", type="primary", key="tb_create_user"):
-                create_user_dialog()
-        with tb2:
-            if st.button("Import Users…", key="tb_import_users"):
-                bulk_import_dialog()
-        with tb3_col:
-            if st.button("Export Users…", key="tb_export_users"):
-                export_users_dialog()
-
-        st.divider()
-
-        left, right = st.columns([1, 2])
-        with left:
-            users = get_all_users()
-            search = st.text_input("Search users", placeholder="Filter by username or name…", key="mu_search")
-            if search:
-                s = search.lower()
-                users = [
-                    u
-                    for u in users
-                    if s in u["username"].lower() or s in f"{u['first_name']} {u['last_name']}`".lower()
-                ]
-            selected = None
-            if users:
-                user_options = {
-                    f"{u['username']} ({u['first_name']} {u['last_name']}) - {u['role_name']}": u for u in users
-                }
-                selected_label = st.selectbox(
-                    "Select a user", options=list(user_options.keys()), key="mu_selected_label"
-                )
-                selected = user_options[selected_label]
-            else:
-                st.info("No users found.")
-
-        with right:
-            if selected is None:
-                st.info("Select a user from the left to view details.")
-            else:
-                st.markdown(
-                    f"**{selected['username']}** — {selected['first_name']} {selected['last_name']} "
-                    f"· {selected['role_name']}"
-                )
-                with SessionLocal() as session:
-                    db_user = session.query(User).filter(User.id == selected["id"]).first()
-                prof_tab, prefs_tab, activity_tab, danger_tab = st.tabs(
-                    ["Profile", "Preferences", "Activity", "Danger Zone"]
-                )
-
-                with prof_tab:
-                    nu_username = st.text_input("Username", value=selected["username"])
-                    nu_first = st.text_input("First Name", value=selected["first_name"])
-                    nu_last = st.text_input("Last Name", value=selected["last_name"])
-                    nu_email = st.text_input("Email", value=selected.get("email") or "")
-                    nu_organization = st.text_input("Organization", value=selected.get("organization") or "")
-                    nu_role_name = st.selectbox(
-                        "Role",
-                        options=role_names,
-                        index=role_names.index(selected["role_name"]) if selected["role_name"] in role_names else 0,
-                    )
-                    theme_options = user_selectable_themes()
-                    current_theme = selected.get("theme", ThemeType.HEALTHELINK.value)
-                    nu_theme = st.selectbox(
-                        "Theme",
-                        options=theme_options,
-                        index=theme_options.index(current_theme) if current_theme in theme_options else 0,
-                    )
-                    prof_btn_cols = st.columns(2)
-                    with prof_btn_cols[0]:
-                        if st.button("Save Profile", key="save_profile", type="primary"):
-                            ok = update_user(
-                                selected["id"],
-                                nu_username,
-                                nu_first,
-                                nu_last,
-                                role_id_by_name.get(nu_role_name),
-                                nu_theme,
-                                email=nu_email.strip() or None,
-                                organization=nu_organization.strip() or None,
-                            )
-                            if ok:
-                                st.success("Profile updated.")
-                                st.rerun()
-                            else:
-                                st.error("Failed to update profile.")
-                    with prof_btn_cols[1]:
-                        if st.button("Set Password…", key="set_password_btn"):
-                            set_password_dialog(selected)
-
-                with prefs_tab:
-                    if db_user:
-                        pref_cols = st.columns(3)
-                        with pref_cols[0]:
-                            p_show_sql = st.checkbox("Show SQL", value=db_user.show_sql)
-                            p_show_table = st.checkbox("Show Table", value=db_user.show_table)
-                            p_plotly = st.checkbox("Show Plotly Code", value=db_user.show_plotly_code)
-                            p_chart = st.checkbox("Show Chart", value=db_user.show_chart)
-                        with pref_cols[1]:
-                            p_history = st.checkbox("Show Question History", value=db_user.show_question_history)
-                            p_summary = st.checkbox("Show Summary", value=db_user.show_summary)
-                            p_voice = st.checkbox("Voice Input", value=db_user.voice_input)
-                            p_speak = st.checkbox("Speak Summary", value=db_user.speak_summary)
-                        with pref_cols[2]:
-                            p_suggested = st.checkbox("Show Suggested", value=db_user.show_suggested)
-                            p_followup = st.checkbox("Show Follow-up", value=db_user.show_followup)
-                            p_elapsed = st.checkbox("Show Elapsed Time", value=db_user.show_elapsed_time)
-                            p_llm = st.checkbox("LLM Fallback", value=db_user.llm_fallback)
-
-                        if st.button("Save Preferences", key="save_prefs", type="primary"):
-                            ok = update_user_preferences(
-                                selected["id"],
-                                show_sql=p_show_sql,
-                                show_table=p_show_table,
-                                show_plotly_code=p_plotly,
-                                show_chart=p_chart,
-                                show_question_history=p_history,
-                                show_summary=p_summary,
-                                voice_input=p_voice,
-                                speak_summary=p_speak,
-                                show_suggested=p_suggested,
-                                show_followup=p_followup,
-                                show_elapsed_time=p_elapsed,
-                                llm_fallback=p_llm,
-                            )
-                            if ok:
-                                st.success("Preferences saved.")
-                            else:
-                                st.error("Failed to save preferences.")
-                    else:
-                        st.info("Preferences unavailable.")
-
-                with activity_tab:
-                    s = stats_map.get(
-                        selected["id"],
-                        {"questions": 0, "charts": 0, "errors": 0, "dataframes": 0, "summaries": 0},
-                    )
-                    m1, m2, m3, m4, m5 = st.columns(5)
-                    m1.metric("Questions", s["questions"])
-                    m2.metric("Charts", s["charts"])
-                    m3.metric("Errors", s["errors"])
-                    m4.metric("DataFrames", s["dataframes"])
-                    m5.metric("Summaries", s["summaries"])
-
-                    range_choice = st.radio(
-                        "Range", options=["7 days", "30 days"], horizontal=True, key="stats_range"
-                    )
-                    days = 7 if range_choice.startswith("7") else 30
-                    daily = get_user_daily_stats(selected["id"], days=days)
-                    if daily:
-                        import plotly.express as px
-
-                        chart_df = pd.DataFrame(daily)
-                        chart_df["date"] = pd.to_datetime(chart_df["date"])
-                        melted = chart_df.melt(
-                            id_vars=["date"],
-                            value_vars=["questions", "charts", "errors", "dataframes", "summaries"],
-                            var_name="metric",
-                            value_name="count",
-                        )
-                        fig = px.line(melted, x="date", y="count", color="metric", markers=True)
-                        fig.update_layout(margin=dict(l=0, r=0, t=10, b=0), legend_title_text="")
-                        st.plotly_chart(fig, use_container_width=True)
-
-                    if st.button(
-                        f"View question audit for {selected['username']} →",
-                        type="primary",
-                        key="activity_audit_deeplink_btn",
-                    ):
-                        st.session_state["audit_trail_pref_user_id"] = selected["id"]
-                        st.switch_page("views/admin_analytics.py")
-
-                with danger_tab:
-                    if st.button("Delete User", type="primary", key="danger_delete_user_btn"):
-                        confirm_destructive(
-                            body_md=(
-                                f"Permanently deletes user **{selected['username']}** "
-                                f"({selected['first_name']} {selected['last_name']}, {selected['role_name']}). "
-                                "Their account, preferences, and message history will be removed."
-                            ),
-                            token="DELETE",
-                            on_confirm=lambda: _delete_and_rerun(selected["id"]),
-                            button_label="Delete User",
-                        )
